requirements.txt
asn1crypto==0.24.0
autopep8==1.4.3
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
Django==2.2
django-filter==2.1.0
django-rest-knox==3.6.0
djangorestframework==3.9.1
httpie==1.0.2
idna==2.8
Markdown==3.0.1
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.3.1
pyOpenSSL==19.0.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
dj-database-url==0.5.0
gunicorn==19.9.0
django-heroku==0.3.1
psycopg2==2.7.7
whitenoise==4.1.2
django-cors-headers==2.5.0
